authoring restrictions

+++ src/routes/cms/index.svelte
@@ -12,9 +12,19 @@
     <title>Dashboard | HOWFEED.BIZ</title>
 </svelte:head>
 
+<script>
+    import FakeTweet from '../../components/FakeTweet.svelte';
+
+    export let user;
+</script>
+
 <div class="content">
-    <h1>HowFeed Publisher Dashboard</h1>
-    <p><a href="/cms/create">Publish a new article</a></p>
-    <p><a href="/cms/update">Edit an existing article</a></p>
-    <p><a href="/cms/delete">Delete an article</a></p>
+    {#if user.author}
+        <h1>HowFeed Publisher Dashboard</h1>
+        <p><a href="/cms/create">Publish a new article</a></p>
+        <p><a href="/cms/update">Edit an existing article</a></p>
+        <p><a href="/cms/delete">Delete an article</a></p>
+    {:else}
+        <FakeTweet message="We're watching you" author="Caltrans HQ" verified likes=0 replies=2 date={new Date(2019, 8, 21)} handle="CaltransHQ" avatar="/ct.jpg" />
+    {/if}
 </div>
